feat: break inbox into some components
