+++ apps/code/src/renderer/features/inbox/components/utils/SignalReportStatusBadge.tsx
