+++ apps/code/src/renderer/features/inbox/components/utils/SignalReportSummaryMarkdown.tsx
