+++ apps/code/src/renderer/features/task-detail/components/TaskDetail.tsx
@@ -12,10 +12,11 @@ import { useWorkspaceEvents } from "@features/workspace/hooks";
 import { useBlurOnEscape } from "@hooks/useBlurOnEscape";
 import { useFileWatcher } from "@hooks/useFileWatcher";
 import { useSetHeaderContent } from "@hooks/useSetHeaderContent";
-import { Box, Flex, Text } from "@radix-ui/themes";
+import { Box, Flex, Text, Tooltip } from "@radix-ui/themes";
 import type { Task } from "@shared/types";
-import { useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { useHotkeys, useHotkeysContext } from "react-hotkeys-hook";
+import { toast } from "sonner";
 import { ExternalAppsOpener } from "./ExternalAppsOpener";
 
 interface TaskDetailProps {
@@ -79,18 +80,33 @@ export function TaskDetail({ task: initialTask }: TaskDetailProps) {
   useBlurOnEscape();
   useWorkspaceEvents(taskId);
 
+  const copyTaskId = useCallback(() => {
+    navigator.clipboard.writeText(taskId);
+    toast.success("Task ID copied");
+  }, [taskId]);
+
   const headerContent = useMemo(
     () => (
       <Flex align="center" justify="between" gap="2" width="100%">
-        <Flex align="center" gap="2" minWidth="0" overflow="hidden">
-          <Text size="1" weight="medium" truncate>
-            {task.title}
-          </Text>
+        <Text size="1" weight="medium" truncate style={{ minWidth: 0 }}>
+          {task.title}
+        </Text>
+        <Flex align="center" gap="2" className="shrink-0">
+          <Tooltip content="Copy task ID">
+            <button
+              type="button"
+              onClick={copyTaskId}
+              className="no-drag cursor-pointer border-0 bg-transparent p-0 font-mono text-[10px] text-gray-9 hover:text-gray-11"
+              style={{ lineHeight: "20px" }}
+            >
+              {taskId}
+            </button>
+          </Tooltip>
+          {openTargetPath && <ExternalAppsOpener targetPath={openTargetPath} />}
         </Flex>
-        {openTargetPath && <ExternalAppsOpener targetPath={openTargetPath} />}
       </Flex>
     ),
-    [task.title, openTargetPath],
+    [task.title, taskId, openTargetPath, copyTaskId],
   );
 
   useSetHeaderContent(headerContent);
